+++ src/main/java/com/programmers/domain/base/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.programmers.base.domain;
+package com.programmers.domain.base;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
