+++ src/main/java/com/programmers/domain/post/infra/PostRepository.java
@@ -1,6 +1,6 @@
-package com.programmers.post.infra;
+package com.programmers.domain.post.infra;
 
-import com.programmers.post.domain.Post;
+import com.programmers.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
