+++ src/main/java/com/programmers/domain/post/ui/dto/PostUpdateDto.java
@@ -1,4 +1,4 @@
-package com.programmers.post.ui.dto;
+package com.programmers.domain.post.ui.dto;
 
 public record PostUpdateDto(String title, String content) {
 }
