+++ src/main/java/com/programmers/domain/user/infra/UserRepository.java
@@ -1,6 +1,6 @@
-package com.programmers.user.infra;
+package com.programmers.domain.user.infra;
 
-import com.programmers.user.domain.User;
+import com.programmers.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
