+++ src/main/java/com/programmers/domain/user/ui/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.programmers.user.ui.dto;
+package com.programmers.domain.user.ui.dto;
 
 public record UserDto(String name, int age, String hobby) {
 }
